SWS_Review_Cloud_Backend_MVP/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
psycopg[binary]==3.2.3
psycopg-pool==3.2.4
pydantic==2.9.2
pydantic-settings==2.5.2
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0,<5.0.0
minio==7.2.12
openpyxl==3.1.5
celery[redis]==5.4.0
redis==5.2.1
python-docx==1.1.2
PyMuPDF==1.25.2
httpx==0.28.1
python-multipart==0.0.22
tqdm==4.66.5
